+++ Editor/Window/Models/CreateFleetResponse.cs
@@ -1,0 +1,23 @@
+﻿using System;
+using AmazonGameLiftPlugin.Core.Shared;
+
+namespace Editor.Window.Models
+{
+    public class GenericResponse : Response
+    {
+        internal GenericResponse()
+        {
+        }
+        
+        public GenericResponse(string errorCode, string errorMessage = null)
+        {
+            if (errorCode is null)
+            {
+                throw new ArgumentNullException(nameof(errorCode));
+            }
+
+            ErrorCode = errorCode;
+            ErrorMessage = errorMessage;
+        }
+    }
+}
