--- Tests/Editor/GameLiftPlugin.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 37e5e48511904a8f8cc6c6e602c3bf20
-timeCreated: 1694448161
