--- Tests/Editor/GameLiftPlugin/AnywherePageTests.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 7a3afd6b27ef436f8f97257481edd0c9
-timeCreated: 1694448176
